pyproject.toml
[project]
name = "rattlesnakeridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-limiter>=4.0.0",
    "flask-session>=0.8.0",
    "openai>=2.7.1",
    "pyyaml>=6.0.3",
]
